Mitsuri Bot/requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
braceexpand==0.1.6
bs4==0.0.1
cachelib==0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudant==2.13.0
colorama==0.4.3
comtypes==1.1.7
cssselect==1.1.0
cycler==0.10.0
dblpy==0.4.0
decorator==4.4.2
demjson==2.2.4
Deprecated==1.2.10
dhooks==1.1.4
discord==1.0.1
discord.py==1.5.0
Django==3.1.2
ffmpeg==1.4
Flask==1.1.2
futures==3.1.1
google-api-core==1.22.2
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
googletrans==3.0.0
goslate==1.5.1
gTTS==2.1.1
gTTS-token==1.1.3
h11==0.9.0
h2==3.2.0
howdoi==2.0.5
hpack==3.0.0
hstspreload==2020.8.8
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
humanize==3.1.0
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
iniconfig==1.1.1
ipykernel==5.3.4
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.6.3
keep==2.9
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lo==0.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
multidict==4.7.6
nose==1.3.7
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDictionary==2.0.1
PyGithub==1.53
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyquery==1.4.1
PySSL==0.0.7
pytest==6.1.1
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.1
pyzmq==19.0.2
ratelimiter==1.2.0.post0
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
simplejson==3.17.2
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
sqlparse==0.3.1
terminaltables==3.1.0
toml==0.10.1
tornado==6.0.4
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
xlrd==1.2.0
xmltodict==0.12.0
yarl==1.5.0
youtube-dl==2020.9.20
